ex_project/requirements.txt
cffi==1.14.5
cryptography==3.4.6
Django==2.2
django-extensions==3.1.1
djangorestframework==3.12.2
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.6
pycparser==2.20
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
drf-yasg